frontend coming along

+++ frontend/src/main/kotlin/chatt/busbus/frontend/GeoLocation.kt
@@ -7,8 +7,8 @@ object GeoLocation {
 
     fun getCurrentPosition(then: (Position) -> Unit) {
         window.navigator.asDynamic().geolocation.getCurrentPosition { pos ->
-            val latitude = (pos.coords.latitude as Float).toDouble()
-            val longitude = (pos.coords.longitude as Float).toDouble()
+            val latitude = (pos.coords.latitude as Double)
+            val longitude = (pos.coords.longitude as Double)
             then(Position(latitude, longitude))
         }
     }
